extract build-menu-filters into a separate file
